+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/orisai/scheduler/compare/1.0.0...v2.x)
+## [Unreleased](https://github.com/orisai/scheduler/compare/2.0.0...v2.x)
+
+### Added
+
+- `ListCommand`
+	- adds `--explain` option to explain whole expression
 
 ## [2.0.0](https://github.com/orisai/scheduler/compare/1.0.0...2.0.0) - 2024-01-26